Added all terms for the FENE P-MP model. Have only implemented this for the cylindrical/axisymmetric case and for the BDF/EXJ version, although it should be fairly similar to adapt the oldB/Giesekus code to implement the semi-implicit scheme too.
Note that there are quite a few unncessary copies going on, so have removed a few, although I have left some in for better readability (and they'd probably be needed for the semi-implicit version).

Still need to integrate this module with the main function (i.e. call the applyElasticStress_FENE_PMP subroutine if the model is selected) - will need to add some input parameters, etc to get this working.